+++ functions/public.ps1
@@ -1,9 +1,77 @@
 
 #region public functions
 
-Function Convert-MySQLiteDB {
-    [cmdletbinding()]
-    [alias('Export-DB')]
+Function Open-MySQLiteDB {
+    [cmdletbinding(SupportsShouldProcess)]
+    [alias("Open-DB")]
+    [OutputType("System.Data.SQLite.SQLiteConnection")]
+
+    Param(
+        [Parameter(Mandatory)]
+        [alias("database")]
+        [string]$Path
+    )
+    Begin {
+        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Starting $($myinvocation.mycommand)"
+
+    } #begin
+
+    Process {
+        $db = resolvedb $Path
+        if ($db.exists) {
+            if ($pscmdlet.shouldprocess($db.path)) {
+                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Opening connection to $Path "
+                opendb $db.path
+            }
+        }
+
+    } #process
+
+    End {
+        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
+
+    } #end
+
+} #close Open-MySQLiteDB
+
+Function Close-MySQLiteDB {
+    [cmdletbinding(SupportsShouldProcess)]
+    [alias("Close-DB")]
+    [OutputType("None", "System.Data.SQLite.SQLiteConnection")]
+
+    Param(
+        [Parameter(Mandatory, HelpMessage = "Enter a connection object", ValueFromPipeline)]
+        [System.Data.SQLite.SQLiteConnection]$Connection,
+        [switch]$Passthru
+    )
+    Begin {
+        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Starting $($myinvocation.mycommand)"
+
+    } #begin
+
+    Process {
+        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Closing connection to $($connection.DataSource)"
+        if ($pscmdlet.ShouldProcess($Connection.DataSource)) {
+            $connection.close()
+            if ($passthru) {
+                $connection
+            }
+
+            $connection.Dispose()
+        }
+
+    } #process
+
+    End {
+        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
+
+    } #end
+
+} #close Close-MySQLiteDB
+
+Function ConvertFrom-MySQLiteDB {
+    [cmdletbinding(DefaultParameterSetName = "table")]
+    [alias('ConvertFrom-DB')]
     [outputtype("Object")]
     Param(
         [Parameter(Position = 0, Mandatory, HelpMessage = "Enter the path to the SQLite database file.", ValueFromPipelineByPropertyName)]
@@ -12,7 +80,17 @@ Function Convert-MySQLiteDB {
         [string]$Path,
         [Parameter(Mandatory, HelpMessage = "Enter the name of the table with data to import")]
         [ValidateNotNullOrEmpty()]
-        [string]$TableName
+        [string]$TableName,
+        [Parameter(Mandatory, HelpMessage = "Enter the name of the property map table", ParameterSetName = "table")]
+        [ValidateNotNullOrEmpty()]
+        [string]$PropertyTable,
+        [Parameter(Mandatory, HelpMessage = "Enter an optional hashtable of property names and types.", ParameterSetName = "hash")]
+        [hashtable]$PropertyMap,
+        [Parameter(HelpMessage = "Enter a typename to insert", ParameterSetName = "hash")]
+        [Parameter(ParameterSetName = "table")]
+        [string]$TypeName,
+        [Parameter(HelpMessage = "Write raw objects to the pipeline.", ParameterSetName = "raw")]
+        [switch]$RawObject
     )
     Begin {
         Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Starting $($myinvocation.mycommand)"
@@ -29,19 +107,63 @@ Function Convert-MySQLiteDB {
             Throw "Failed to find database file $($file.path)"
         }
         #verify table exists
-        $tables = Get-mySQLiteTable -connection $connection -KeepAlive
+        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Verifying table $tablename"
+        $tables = Get-MySQLiteTable -connection $connection -KeepAlive
         if ($tables.name -contains $tablename) {
             $query = "Select * from $tablename"
             Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Importing from table $Tablename"
-            $raw = Invoke-mySQLiteQuery -connection $connection -Query $query -As object -KeepAlive
+            Try {
+                $raw = Invoke-MySQLiteQuery -connection $connection -Query $query -As object -KeepAlive -ErrorAction stop
+            }
+            Catch {
+                Write-Warning $_.exception.message
+                closedb $connection
+                Throw $_
+                #bail out
+                return
+            }
             Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Found $($raw.count) items"
 
-            #use propertymap table if it exists
-            if ($tables.name -contains "PropertyMap") {
-                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Getting property map"
-                $map = Invoke-mySQLiteQuery -Connection $connection -Query "Select * from propertymap" -KeepAlive -as Hashtable
+            <#
+                find a mapping table using this priority list
+                1. PropertyMap parameter
+                2. A table called PropertyMap_tablename
+
+                if nothing found then write a default custom object
+            #>
+            switch ($pscmdlet.ParameterSetName) {
+                "hash" {
+                    Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] User specified property map"
+                    $map = $PropertyMap
+                    If ($TypeName) {
+                        $oTypename = $TypeName
+                    }
+                }
+                "table" {
+                    Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Getting property map from $PropertyTable"
+                    $map = Invoke-MySQLiteQuery -Connection $connection -Query "Select * from $propertytable" -KeepAlive -as Hashtable
+                    if ($typename) {
+                        $oTypename = $TypeName
+                    }
+                    elseif ($PropertyTable -match "_") {
+                        #get the typename from the property table name
+                        $oTypename = $PropertyTable.split("_", 2)[1].replace("_", ".")
+                    }
+
+                }
+                "raw" {
+                    Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Writing raw objects to the pipeline"
+                    $raw
+                }
+            }
+
+            if ($map) {
                 foreach ($item in $raw) {
                     $tmpHash = [ordered]@{}
+                    if ($oTypename) {
+                        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Inserting typename $oTypename"
+                        $tmpHash.Add("PSTypename", $oTypename)
+                    }
                     foreach ($key in $map.keys) {
                         Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] ...$key"
                         $name = $key
@@ -53,16 +175,12 @@ Function Convert-MySQLiteDB {
                             $v = $item.$key
                         }
                         $value = $v -as $($($map[$key] -as [type]))
-                        $tmpHash.Add($name,$value)
-                    }
+                        $tmpHash.Add($name, $value)
+                    } #foreach key
                     New-Object -typename PSObject -property $tmpHash
-                }
-            }
-            else {
-                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Writing raw objects to the pipeline"
-                $raw
-            }
-        }
+                } #foreach item
+            } #if $map
+        } #if table found
         else {
             Write-Warning "Failed to find a table called $Tablename in $($file.path)"
         }
@@ -70,14 +188,15 @@ Function Convert-MySQLiteDB {
     } #process
 
     End {
-        closedb $connection
+        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Closing database connection"
+        closedb -connection $connection
         Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
 
     } #end
 
-} #close Convert-MySQLiteDB
+} #close ConvertFrom-MySQLiteDB
 
-Function Get-mySQLiteTable {
+Function Get-MySQLiteTable {
     [cmdletbinding(DefaultParameterSetName = "file")]
     [Alias("gtb", "Get-DBTable")]
     [outputtype("PSCustomObject")]
@@ -90,18 +209,23 @@ Function Get-mySQLiteTable {
         [Parameter(Position = 0, HelpMessage = "Specify an existing open database connection.", ParameterSetName = "connection")]
         [System.Data.SQLite.SQLiteConnection]$Connection,
         [Parameter(HelpMessage = "Do not close the connection.", ParameterSetName = "connection")]
-        [switch]$KeepAlive
+        [switch]$KeepAlive,
+        [switch]$Detail
     )
     Begin {
         Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Starting $($myinvocation.mycommand)"
     } #begin
 
     Process {
+        $iqParams = @{
+            query = "Select Name from sqlite_master where type='table'"
+        }
         if ($pscmdlet.ParameterSetName -eq 'file') {
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Getting tables from $Path"
             $db = resolvedb -Path $path
+            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Getting tables from $($db.Path)"
             if ($db.exists) {
-                Invoke-mySQLiteQuery -Path $db.path -query "Select Name from sqlite_master where type='table'"
+                $iqParams.Add("Path", $db.path)
+                $source = $db.path
             }
             else {
                 Write-Warning "Failed to find $($db.path)"
@@ -110,7 +234,43 @@ Function Get-mySQLiteTable {
         else {
             Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Using an existing connection $($connection.ConnectionString)"
             Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] KeepAlive is $KeepAlive"
-            Invoke-mySQLiteQuery -connection $Connection -query "Select Name from sqlite_master where type='table'" -KeepAlive:$KeepAlive
+            $iqParams.Add("Connection", $Connection)
+            $iqParams.Add("KeepAlive", $KeepAlive)
+            #parse out the path from the connection
+            $source = $connection.ConnectionString.split("=", 2)[1].split(";")[0]
+
+        }
+        $tablenames = Invoke-MySQLiteQuery @iqParams
+        if ($tablenames) {
+
+            if ($Detail) {
+                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Getting table details"
+                foreach ($item in $Tablenames.name) {
+                    $iqParams.query = "PRAGMA table_info($item)"
+                    Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] $($iqparams.query)"
+                    $details = Invoke-MySQLiteQuery @iqParams
+                    foreach ($tbl in $details) {
+                        [pscustomobject]@{
+                            PSTypename  = "MySQLiteTableDetail"
+                            Source      = $Source
+                            Table       = $item
+                            ColumnIndex = $tbl.cid
+                            ColumnName  = $tbl.name
+                            ColumnType  = $tbl.type
+                        }
+                    } #foreach tbl
+                } #foreach item
+            }
+            else {
+                [pscustomobject]@{
+                    Source = $source
+                    Name   = $Tablenames.name
+                }
+
+            }
+        }
+        else {
+            Write-Warning "No tables found in $source"
         }
     } #process
 
@@ -122,9 +282,9 @@ Function Get-mySQLiteTable {
         Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
     } #end
 
-} #close Get-mySQLiteTable
+} #close Get-MySQLiteTable
 
-Function ConvertTo-mySQLiteDB {
+Function ConvertTo-MySQLiteDB {
     [cmdletbinding(SupportsShouldProcess)]
     [alias("todb", 'Convert-DB')]
     [outputtype("None")]
@@ -138,15 +298,18 @@ Function ConvertTo-mySQLiteDB {
         [Parameter(Mandatory, HelpMessage = "Enter the name of the new table. Table names are technically case-sensitive.")]
         [ValidateNotNullOrEmpty()]
         [string]$TableName,
-        [switch]$Append
+        [Parameter(HelpMessage = "Enter a typename for your converted objects. If you don't specify one, it will be auto-detected.")]
+        [ValidatePattern("^\w+$")]
+        [string]$TypeName,
+        [switch]$Append,
+        [switch]$Force
     )
     Begin {
         Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Starting $($myinvocation.mycommand)"
-
+        $file = resolvedb -Path $path
         if ($Append) {
-            $file = resolvedb -Path $path
             if ($file.exists) {
-                $db = opendb -Path $file.path
+                $connection = opendb -Path $file.path
             }
             else {
                 Throw "Failed to find database file $($file.path)"
@@ -154,14 +317,23 @@ Function ConvertTo-mySQLiteDB {
         }
         else {
             $newParams = @{
-                Path     = $executioncontext.sessionstate.path.GetUnresolvedProviderPathFromPSPath($path)
-                Force    = $True
-                Passthru = $True
+                Path        = $file.path
+                Force       = $Force
+                Passthru    = $True
+                ErrorAction = "Stop"
             }
             Try {
                 if ($PSCmdlet.ShouldProcess($Path, "Create Database")) {
-                    $db = New-mySQLiteDB @newParams
+                    Try {
+                        $db = New-MySQLiteDB @newParams
+                    }
+                    Catch {
+                        Throw $_
+                        #bail out
+                        return
+                    }
                     Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Created database at $($db.fullname)"
+                    $connection = opendb $db.fullname
                 }
             }
             Catch {
@@ -172,28 +344,55 @@ Function ConvertTo-mySQLiteDB {
     } #begin
 
     Process {
+        $iqParams = @{
+            Connection = $Connection
+            KeepAlive  = $True
+            Query      = $null
+        }
 
         foreach ($object in $Inputobject) {
 
             if ($TableExists) {
                 Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Adding object to the table"
-                $query = buildquery -InputObject $object -Tablename $TableName
+                $iqParams.query = buildquery -InputObject $object -Tablename $TableName
                 if ($pscmdlet.ShouldProcess("object", "Add to table $Tablename")) {
-                    Invoke-mySQLiteQuery -Path $Path -Query $query
+                    Invoke-MySQLiteQuery @iqParams
                 }
             }
             else {
                 # http://www.sqlitetutorial.net/sqlite-data-types/
+                # https://www.sqlite.org/datatype3.html
                 #convert types as necessary. Table types can be Text, Int, Real or Blob
-                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Creating property map table"
                 if ($PSCmdlet.ShouldProcess("PropertyMap", "Create Table")) {
-                    New-mySQLiteDBTable -Path $path -TableName PropertyMap -ColumnNames $object.psobject.properties.name
+                    $object.psobject.properties |
+                        foreach-object -begin {
+                        $prop = [ordered]@{}
+                    } -process {
+                        $prop.Add($_.Name, "Text")
+                    }
+
+                    if ($Typename) {
+                        $name = "propertymap_{0}" -f ($typename.tolower())
+                    }
+                    else {
+                        $name = "propertymap_{0}" -f ($object.psobject.typenames[0].replace(".", "_"))
+                    }
+                    Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Creating property map table $name"
+                    $prop | Out-String | Write-Verbose
+                    $newTblParams = @{
+                        Connection       = $Connection
+                        KeepAlive        = $True
+                        TableName        = $Name
+                        ColumnProperties = $prop
+                    }
+                    New-MySQLiteDBTable @newTblParams
                 }
+
                 $names = $object.psobject.properties.name -join ","
                 $values = $object.psobject.properties.TypeNameofValue -join "','"
-                $query = "Insert Into PropertyMap ($names) values ('$values')"
+                $iqParams.query = "Insert Into $Name ($names) values ('$values')"
                 if ($PSCmdlet.ShouldProcess($query, "Run query")) {
-                    Invoke-mySQLiteQuery -Path $Path -Query $query
+                    Invoke-MySQLiteQuery @iqParams
                 }
 
                 Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Creating property hashtable"
@@ -202,29 +401,32 @@ Function ConvertTo-mySQLiteDB {
                 $thash = [ordered]@{}
                 Foreach ($prop in $properties) {
                     Switch -Regex ($prop.TypeNameofValue) {
-                        "Int32$"            {$sqltype = "Int" }
-                        "Int64$"            {$sqltype = "Real"}
-                        "^System.Double$"   {$sqltype = "Real"}
-                        "^System.DateTime"  {$sqltype = "Text"}
-                        "^System.String$"   {$sqltype = "Text"}
-                        "^System.Boolean$"  {$sqltype = "Text"}
+                        "Int32$" {$sqltype = "Int" }
+                        "Int64$" {$sqltype = "Real"}
+                        "^System.Double$" {$sqltype = "Real"}
+                        "^System.DateTime" {$sqltype = "Text"}
+                        "^System.String$" {$sqltype = "Text"}
+                        "^System.Boolean$" {$sqltype = "Int"}
                         default {
                             $sqltype = "Blob"
                         }
                     } #switch
                     $thash.Add($prop.Name, $sqltype)
                 } #foreach prop
+
                 Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Creating table: $Tablename"
                 if ($pscmdlet.ShouldProcess($tablename, "Create table")) {
-                    New-mySQLiteDBTable -Path $path -TableName $TableName -columnProperties $thash
+                    $newTblParams.ColumnProperties = $thash
+                    $newtblParams.tablename = $Tablename
+                    New-MySQLiteDBTable @newtblParams
                 }
 
                 Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Inserting the first object into the table"
                 #insert the first object into the new table
-                $query = buildquery -InputObject $object -Tablename $TableName
+                $iqParams.query = buildquery -InputObject $object -Tablename $TableName
 
                 if ($pscmdlet.ShouldProcess("object", "Insert first object")) {
-                    Invoke-mySQLiteQuery -Path $Path -Query $query
+                    Invoke-MySQLiteQuery @iqParams
                 }
                 $TableExists = $True
             }
@@ -232,12 +434,15 @@ Function ConvertTo-mySQLiteDB {
     } #process
 
     End {
+        if ($connection.State -eq "open") {
+            closedb -connection $connection
+        }
         Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
     } #end
 
-} #close Convert-mySQLiteDB
+} #close Convert-MySQLiteDB
 
-Function New-mySQLiteDB {
+Function New-MySQLiteDB {
     [cmdletbinding(SupportsShouldProcess)]
     [alias("New-DB", "ndb")]
     [outputtype("None", "System.IO.Fileinfo")]
@@ -260,6 +465,7 @@ Function New-mySQLiteDB {
 
         If (($db.exists) -AND (-not $Force)) {
             Write-Warning "The database file $path exists. Use -Force to overwrite the file."
+            throw "The database file $path exists."
         }
         else {
             If (($db.exists) -AND $Force) {
@@ -287,17 +493,17 @@ Function New-mySQLiteDB {
 
             [string]$query = "CREATE TABLE Metadata (Author TEXT,Created TEXT,Computername TEXT,Comment TEXT);"
 
-            $cmd = $connection.CreateCommand()
-            $cmd.CommandText = $query
             if ($pscmdlet.ShouldProcess($query)) {
+                $cmd = $connection.CreateCommand()
+                $cmd.CommandText = $query
                 [void]$cmd.ExecuteNonQuery()
             }
 
             $query = "Insert Into Metadata (Author,Created,Computername,Comment) Values ('$($meta.author)','$($meta.created)','$($meta.computer)','$($meta.comment)')"
 
             Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] $query"
-            $cmd.CommandText = $query
             if ($pscmdlet.ShouldProcess($query)) {
+                $cmd.CommandText = $query
                 [void]$cmd.ExecuteNonQuery()
             }
         }
@@ -318,41 +524,66 @@ Function New-mySQLiteDB {
 
     } #end
 
-} #new-mysqlitedb
+} #new-MySQLitedb
 
-Function New-mySQLiteDBTable {
-    [cmdletbinding(SupportsShouldProcess, DefaultParameterSetName = "typed")]
+Function New-MySQLiteDBTable {
+    [cmdletbinding(SupportsShouldProcess, DefaultParameterSetName = "filetyped")]
     [alias("New-DBTable", "ndbt")]
     [outputtype("None")]
     Param(
-        [Parameter(Position = 0, Mandatory, HelpMessage = "Enter the path to the SQLite database file.", ValueFromPipelineByPropertyName)]
+        [Parameter(Position = 0, Mandatory, HelpMessage = "Enter the path to the SQLite database file.", ValueFromPipelineByPropertyName, ParameterSetName = "filetyped")]
+        [Parameter(ParameterSetName = "filenamed")]
         [Alias("fullname", "database")]
         [ValidateNotNullOrEmpty()]
         [string]$Path,
+
+        [Parameter(HelpMessage = "Specify an existing open database connection.", ParameterSetName = "cnxtyped")]
+        [Parameter(ParameterSetName = "cnxnamed")]
+        [ValidateNotNullOrEmpty()]
+        [System.Data.SQLite.SQLiteConnection]$Connection,
+
         [Parameter(Mandatory, HelpMessage = "Enter the name of the new table. Table names are technically case-sensitive.")]
         [ValidateNotNullOrEmpty()]
         [string]$TableName,
-        [parameter(Mandatory, HelpMessage = "Enter an ordered hashtable of column definitions", ParameterSetName = "typed")]
+
+        [parameter( HelpMessage = "Enter an ordered hashtable of column definitions", ParameterSetName = "filetyped")]
+        [Parameter(ParameterSetName = "cnxtyped")]
+        [ValidateNotNullOrEmpty()]
         [System.Collections.Specialized.OrderedDictionary]$ColumnProperties,
-        [parameter(Mandatory, HelpMessage = "Enter an array of column names.", ParameterSetName = "untyped")]
+
+        [parameter( HelpMessage = "Enter an array of column names.", ParameterSetName = "cnxnamed")]
+        [Parameter(ParameterSetName = "filenamed")]
+        [ValidateNotNullOrEmpty()]
         [string[]]$ColumnNames,
+
         [Parameter(HelpMessage = "Overwrite an existing table. This could result in data loss.")]
-        [switch]$Force
+        [switch]$Force,
+
+        [Parameter(HelpMessage = "Keep an existing connection open.", ParameterSetName = "cnxtyped")]
+        [Parameter(ParameterSetName = "cnxnamed")]
+        [switch]$KeepAlive
     )
 
     Begin {
         Write-Verbose "[$((Get-Date).TimeofDay) BEGIN] Starting $($myinvocation.mycommand)"
     } #begin
     Process {
-        $db = resolvedb -path $Path
 
-        If ($db.exists) {
-            $connection = opendb $db.path
+        if ($Path) {
+
+            $db = resolvedb -path $Path
+
+            If ($db.exists) {
+                $connection = opendb $db.path
+            }
+            else {
+                Write-Warning "Cannot find the database file $($db.path)."
+            }
+            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Processing database: $($db.Path)"
         }
         else {
-            Write-Warning "Cannot find the database file $($db.path)."
+            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Using an existing connection"
         }
-        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Processing database: $($db.Path)"
         if ($connection.state -eq 'Open' -OR $PSBoundparameters.ContainsKey("WhatIf")) {
 
             $cmd = $connection.CreateCommand()
@@ -388,7 +619,7 @@ Function New-mySQLiteDBTable {
                 Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Adding Table: $Tablename"
                 [string]$query = "CREATE TABLE $tablename "
 
-                if ($pscmdlet.ParameterSetName -eq 'typed') {
+                if ($pscmdlet.ParameterSetName -match 'typed') {
                     $keys = $ColumnProperties.Keys
                     $primary = $keys | Select-Object -first 1`
                     $primaryType = $ColumnProperties.item($Primary)
@@ -409,6 +640,7 @@ Function New-mySQLiteDBTable {
 
                 $cmd = $connection.CreateCommand()
                 $cmd.CommandText = $query
+                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] $query"
                 if ($pscmdlet.ShouldProcess($query)) {
                     [void]$cmd.ExecuteNonQuery()
                 }
@@ -419,19 +651,18 @@ Function New-mySQLiteDBTable {
         }
     } #process
     End {
-        if ($connection.state -eq 'Open') {
+        if ($connection.state -eq 'Open' -AND (-Not $KeepAlive)) {
             Write-Verbose "[$((Get-Date).TimeofDay) END    ] Closing database connection"
-            $connection.close()
-            $connection.Dispose()
+            closedb -connection $connection -cmd $cmd
         }
         Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
     } #end
-} #new-mysqlitedbtable
+} #new-MySQLitedbtable
 
-Function Invoke-mySQLiteQuery {
+Function Invoke-MySQLiteQuery {
     [cmdletbinding(SupportsShouldProcess, DefaultParameterSetName = "file")]
     [alias("iq")]
-    [outputtype("None", "PSCustomObject", "System.Data.Datatable")]
+    [outputtype("None", "PSCustomObject", "System.Data.Datatable", "Hashtable")]
     Param(
         [Parameter(Position = 0, Mandatory, HelpMessage = "Enter the path to the SQLite database file.", ParameterSetName = "file", ValueFromPipelineByPropertyName)]
         [Alias("fullname", "database")]
@@ -441,10 +672,10 @@ Function Invoke-mySQLiteQuery {
         [System.Data.SQLite.SQLiteConnection]$Connection,
         [Parameter(Position = 1, Mandatory, HelpMessage = "Enter a SQL query string")]
         [string]$Query,
-        [Parameter(HelpMessage = "Keep the connection alive.",ParameterSetName = "connection")]
+        [Parameter(HelpMessage = "Keep the connection alive.", ParameterSetName = "connection")]
         [switch]$KeepAlive,
         [Parameter(HelpMessage = "Write the results of a Select query in the specified format")]
-        [ValidateSet("Object","Datatable","Hashtable")]
+        [ValidateSet("Object", "Datatable", "Hashtable")]
         [string]$As = "object"
     )
     Begin {
@@ -457,10 +688,10 @@ Function Invoke-mySQLiteQuery {
             $file = resolvedb -path $path
 
             If ($file.exists) {
-                $connection = opendb -Path $path
+                $connection = opendb -Path $file.path
             }
             else {
-                Write-Warning "Cannot find the database file $path."
+                Write-Warning "Cannot find the database file $file.path."
             }
         }
         else {
@@ -477,7 +708,8 @@ Function Invoke-mySQLiteQuery {
             if ($pscmdlet.ShouldProcess($query)) {
                 #determine what method to invoke based on the query
                 Switch -regex ($query) {
-                    "^Select (\w+|\*)|(@@\w+ AS)" {
+                    "^([Ss]elect (\w+|\*)|(@@\w+ AS))|([Pp]ragma \w+)" {
+                        #   "^Select (\w+|\*)|(@@\w+ AS)" {
                         if ($As -eq "datatable") {
                             Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Datatable"
                             $ds = New-Object System.Data.DataSet
@@ -537,6 +769,71 @@ Function Invoke-mySQLiteQuery {
         }
         Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
     } #end
-} #close invoke-mysqlitequery
+} #close invoke-MySQLitequery
+
+
+Function Get-MySQLiteDB {
+    [cmdletbinding()]
+    [alias('Get-DB')]
+    [OutputType('MySQLiteDB')]
+    Param(
+        [Parameter(Position = 0, Mandatory, HelpMessage = "Enter the path to the SQLite database file.", ValueFromPipelineByPropertyName)]
+        [Alias("fullname", "database")]
+        [ValidateNotNullOrEmpty()]
+        [string]$Path
+    )
+    Begin {
+        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Starting $($myinvocation.mycommand)"
+
+    } #begin
+
+    Process {
+
+        $file = resolvedb -path $path
+
+        If ($file.exists) {
+            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Getting database information from $($file.path)"
+            $thisdb = Get-Item -path $file.path
+
+            $connection = opendb -Path $file.path
+
+            $tables = Get-MySQLiteTable -Connection $connection -KeepAlive | Select-Object -ExpandProperty Name
+            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Found $($tables.count) Tables"
+
+            $pgsize = (Invoke-MySQLiteQuery -connection $connection -query "PRAGMA page_size" -KeepAlive ).page_size
+            $pgcount = (Invoke-MySQLiteQuery -connection $connection -query "PRAGMA page_count" -KeepAlive ).page_count
+            $encoding = (Invoke-MySQLiteQuery -connection $connection -query "PRAGMA encoding" -KeepAlive).encoding
+
+            [pscustomobject]@{
+                PSTypename    = "MySQLiteDB"
+                DatabaseName  = $connection.Database
+                Tables        = $tables
+                PageSize      = $pgsize
+                PageCount     = $pgcount
+                Encoding      = $encoding
+                FileName      = $thisdb.name
+                Path          = $File.path
+                Size          = $thisdb.length
+                Created       = $thisdb.Creationtime
+                Modified      = $thisdb.LastWriteTime
+                Age           = (Get-Date) - $thisdb.LastWriteTime
+                SQLiteVersion = $connection.serverversion
+            }
+        }
+        else {
+            Write-Warning "Cannot find the database file $path."
+        }
+
+    } #process
+
+    End {
+        if ($connection.open) {
+            Write-Verbose "[$((Get-Date).TimeofDay) END    ] Closing database connection"
+            closedb -connection $connection
+        }
+        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
+    } #end
+
+} #close Get-MySQLiteDB
 
 #endregion